% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulation_ratio.R
\name{accumulation_ratio}
\alias{accumulation_ratio}
\title{Calculate accumulation ratio
This is the ratio of drug concentration or AUC at steady state over concentrations after single dose}
\usage{
accumulation_ratio(kel = NULL, halflife = NULL, tau = 24)
}
\arguments{
\item{kel}{drug elimination rate}

\item{halflife}{halflife. Either `kel` or `halflife` is required.}

\item{tau}{dosing interval}
}
\description{
Calculate accumulation ratio
This is the ratio of drug concentration or AUC at steady state over concentrations after single dose
}
\examples{
accumulation_ratio(halflife = 24, tau = 24)
accumulation_ratio(kel = 0.08, tau = 12)
}
