% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ruv.R
\name{add_ruv}
\alias{add_ruv}
\title{Add residual variability to data}
\usage{
add_ruv(x, ruv = list())
}
\arguments{
\item{x}{data}

\item{ruv}{list with arguments prop, add, exp}
}
\description{
Add residual variability to data
}
