% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aki_stage.R
\name{calc_aki_stage}
\alias{calc_aki_stage}
\title{Calculate AKI stage}
\usage{
calc_aki_stage(scr = NULL, t = NULL, age = NULL, egfr = NULL,
  recursive = TRUE, verbose = FALSE)
}
\arguments{
\item{scr}{serum creatinine in mg/dL. Use `convert_creat()` to convert from mmol/L.}

\item{age}{age in years}

\item{egfr}{eGFR in ml/min/1.73m^2}

\item{recursive}{recursive (if `FALSE` will only take last observation into account)}

\item{verbose}{verbose (`TRUE` or `FALSE`)}

\item{times}{creatinine sample times in hours}
}
\description{
Calculate stage of kidney injury, following definition from KDIGO (http://www.kdigo.org/clinical_practice_guidelines/pdf/KDIGO%20AKI%20Guideline.pdf).
Notes:
- will only take serum creatinine into account
- full serum creatinine history (at least up to 7 days prior) has to be provided.
- `egfr` and `age` are optional arguments, will only be taken into account if supplied, no warning is shown if not supplied.
}
\details{
AKI definition:
AKI is defined as any of the following (
 - increase in SCr by >= 0.3 mg/dL within 48 hours
 - increase in SCr to >=1.5x baseline (baseline in prior 7 days)
 - urine volume < 0.5 ml/kg/h for 6 hours

AKI Staging:
Stage 1: SCr 1.5–1.9 times baseline OR >=0.3 mg/dl (>=26.5 mmol/l) increase
Stage 2: SCr 2.0–2.9 times baseline
Stage 3: SCr 3.0 times baseline OR Initiation of RRT, OR, in patients < 18 y.o., decrease in GFR to <35 ml/min per 1.73m2
}
\examples{
calc_aki_stage(scr = c(0.7, 0.9, 1.8, 1.5), times = c(0, 40, 100, 130), sex = "male", age = 40)
}
