% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bmi.R
\name{calc_bmi}
\alias{calc_bmi}
\title{Calculate BMI in kg/m2}
\usage{
calc_bmi(weight, height)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}
}
\value{
value of BMI in kg/m2
}
\description{
Calculate BMI in kg/m2
}
\examples{
calc_bmi(weight = 70, height = 160)
}
