% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bsa.R
\name{calc_bsa}
\alias{calc_bsa}
\title{Calculate body surface area}
\usage{
calc_bsa(weight = NULL, height = NULL, method = "dubois")
}
\arguments{
\item{weight}{weight}

\item{height}{height}

\item{method}{estimation method, choose from `dubois`, `mosteller`, `haycock`, `gehan_george`, `boyd`}
}
\description{
Get an estimate of body-surface area based on weight and height
}
