% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cefepime_neurotoxicity.R
\name{calc_cefepime_neurotoxicity}
\alias{calc_cefepime_neurotoxicity}
\title{Calculate Cefepime-associated Neurotoxicity}
\usage{
calc_cefepime_neurotoxicity(trough_concentration)
}
\arguments{
\item{trough_concentration}{Cefepime plasma trough concentration (mg/L).}
}
\value{
Probability of cefepime-associated neurotoxicity.
}
\description{
Calculate probability of cefepime-associated neurotoxicity as a function of
cefepime plasma trough concentrations. This model was based on a data set
of adult patients receiving thrice-daily cefepime and may not correspond to
other dosing intervals or patient populations.
}
\note{
The intercept and slope used in this function match Figure 2 in the cited
paper, but differ slightly from the values provided in the text. This 
interpretation was confirmed via correspondence with the authors.
}
\examples{
calc_cefepime_neurotoxicity(trough_concentration = c(13.7, 17.8))
}
\references{
\href{https://doi.org/10.1016/j.cmi.2019.06.028}{
Boschung-Pasquier et al., Clinical Microbiology and Infection (2020)}
}
