% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dose_auc_cmax_cmin.R
\name{calc_dose_auc_cmax_cmin}
\alias{calc_dose_auc_cmax_cmin}
\title{Calculate dose based on a given AUC24, Cmax, and Cmin, assuming 1-compartment model}
\usage{
calc_dose_auc_cmax_cmin(target = list(), parameters = list(),
  interval = 24, t_inf = 1, round_interval = TRUE)
}
\arguments{
\item{target}{list of `auc`, `cmax`, and `cmin`}

\item{parameters}{list of `CL` and `V`}

\item{interval}{dosing interval}

\item{t_inf}{infusion time}

\item{round_interval}{round interval to nearest nominal interval?}
}
\description{
Calculate dose based on a given AUC24, Cmax, and Cmin, assuming 1-compartment model
}
