% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dosing_weight.R
\name{calc_dosing_weight}
\alias{calc_dosing_weight}
\title{Calculate commonly used "dosing weight", based on total (TBW), ideal (IBW), or adjusted (ABW) body weight
This is derived using following:
- In principle, use IBW
- If total body weight (TBW) > 1.2*IBW, then use ABW
- If TBW < IBW, use TBW}
\usage{
calc_dosing_weight(weight, height, age, sex, verbose = TRUE, ...)
}
\arguments{
\item{weight}{weight}

\item{height}{height}

\item{age}{age}

\item{sex}{sex}

\item{...}{pased to `calc_abw()` function}
}
\description{
Calculate commonly used "dosing weight", based on total (TBW), ideal (IBW), or adjusted (ABW) body weight
This is derived using following:
- In principle, use IBW
- If total body weight (TBW) > 1.2*IBW, then use ABW
- If TBW < IBW, use TBW
}
\examples{
calc_dosing_weight(weight = 50, height = 170, sex = "female", age = 50)
}
