% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egfr.R
\name{calc_egfr}
\alias{calc_egfr}
\title{Calculate eGFR}
\usage{
calc_egfr(
  method = "cockcroft_gault",
  sex = NULL,
  age = NULL,
  scr = NULL,
  scr_unit = NULL,
  race = "other",
  weight = NULL,
  height = NULL,
  bsa = NULL,
  preterm = FALSE,
  ckd = FALSE,
  times = NULL,
  bsa_method = "dubois",
  relative = NULL,
  unit_out = "mL/min",
  verbose = TRUE,
  min_value = NULL,
  max_value = NULL,
  ...
)
}
\arguments{
\item{method}{eGFR estimation method, choose from `cockcroft_gault`, `cockcroft_gault_ideal`, `cockcroft_gault_adjusted`, `cockcroft_gault_adaptive `mdrd`, `ckd_epi`, malmo_lund_revised`, `schwartz`, `jelliffe`, `jellife_unstable`, `wright`}

\item{sex}{sex}

\item{age}{age}

\item{scr}{serum creatinine (mg/dL)}

\item{scr_unit, }{`mg/dL` or `micromol/L` (==`umol/L`)}

\item{race}{`black` or `other`}

\item{weight}{weight}

\item{height}{height, only relevant when converting to/from BSA-relative unit}

\item{bsa}{body surface area}

\item{preterm}{is patient preterm?}

\item{ckd}{chronic kidney disease? (Schwartz equations only)}

\item{times}{vector of sampling times (in days!) for creatinine (only used in Jelliffe equation for unstable patients)}

\item{bsa_method}{BSA estimation method, see `bsa()` for details}

\item{relative}{`TRUE`/`FALSE`. Report eGFR as per 1.73 m2? Requires BSA if re-calculation required. If `NULL` (=default), will choose value typical for `method`.}

\item{unit_out}{`ml/min` (default), `L/hr`, or `mL/hr`}

\item{verbose}{verbocity, show guidance and warnings. `TRUE` by default}

\item{min_value}{minimum value (`NULL` by default). The cap is applied in the same unit as the `unit_out`.}

\item{max_value}{maximum value (`NULL` by default). The cap is applied in the same unit as the `unit_out`.}

\item{...}{arguments passed on}
}
\description{
Calculate the estimated glomerulal filtration rate (an estimate of renal function) based on measured serum creatinine using one of the following approaches:
\itemize{
  \item Cockcroft-Gault (using weight, ideal body weight, or adjusted body weight)
  \item C-G spinal cord injury
  \item Revised Lund-Malmo
  \item Modification of Diet in Renal Disease study (MDRD)
  \item Schwartz
  \item Schwartz revised / bedside
  \item Jelliffe
  \item Jelliffe for unstable renal function
  \item Wright
}
Equations for estimation of eGFR from Cystatin C concentrations are available from the `calc_egfr_cystatin()` function.
}
\examples{
calc_egfr(sex = "male", age = 50, scr = 1.1, weight = 70)
calc_egfr(sex = "male", age = 50, scr = 1.1, weight = 70, unit_out = "L/hr")
calc_egfr(sex = "male", age = 50, scr = 1.1, weight = 70, bsa = 1.8, method = "ckd_epi")
calc_egfr(sex = "male", age = 50, scr = c(1.1, 0.8),
  weight = 70, height = 170, method = "jelliffe")
calc_egfr(sex = "male", age = 50, scr = c(1.1, 0.8),
  weight = 70, height = 170, method = "jelliffe_unstable")
calc_egfr(sex = "male", age = 50, scr = 1.1,
  weight = 70, bsa = 1.6, method = "malmo_lund_rev", relative = FALSE)
}
\references{
\itemize{
  \item Cockcroft-Gault: \href{http://www.ncbi.nlm.nih.gov/pubmed/1244564}{Cockcroft & Gault, Nephron (1976)}
  \item Cockcroft-Gault for spinal cord injury: \href{https://www.ncbi.nlm.nih.gov/pubmed/6835689}{Mirahmadi et al., Paraplegia (1983)}
  \item Revised Lund-Malmo: \href{http://www.ncbi.nlm.nih.gov/pubmed/24334413}{Nyman et al., Clinical Chemistry and Laboratory Medicine (2014)}
  \item MDRD: \href{https://www.ncbi.nlm.nih.gov/pubmed/10075613}{Level et al., Annals of Internal Medicine}
  \item Schwartz: \href{https://www.ncbi.nlm.nih.gov/pubmed/951142}{Schwartz et al., Pediatrics (1976)}
  \item Schwartz revised / bedside: \href{https://www.ncbi.nlm.nih.gov/pubmed/19158356}{Schwartz et al., Journal of the American Society of Nephrology (2009)}
  \item Jelliffe: \href{https://www.ncbi.nlm.nih.gov/pubmed/4748282}{Jelliffe, Annals of Internal Medicine (1973)}
  \item Jelliffe for unstable renal function: \href{https://www.ncbi.nlm.nih.gov/pubmed/4748282}{Jelliffe, American Journal of Nephrology (2002)}
  \item Wright: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2363765/}{Wright et al., British Journal of Cancer (2001)}
}
}
