% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kel_single_tdm.R
\name{calc_kel_single_tdm}
\alias{calc_kel_single_tdm}
\title{Calculate elimination rate when given a single TDM sample}
\usage{
calc_kel_single_tdm(
  dose = 1000,
  V = 50,
  t = 10,
  dv = 10,
  tau = 12,
  t_inf = 1,
  kel_init = 0.1,
  n_iter = 25,
  learn_rate = 0.2
)
}
\arguments{
\item{dose}{dose amount}

\item{V}{volume of distribution}

\item{t}{time or time after dose}

\item{dv}{observed value}

\item{tau}{dosing interval}

\item{t_inf}{infusion time}

\item{kel_init}{estimate of elimination rate}

\item{n_iter}{number of iterations to improve estimate of elimination rate}

\item{learn_rate}{default is 0.2}
}
\description{
Using iterative k_el calculation, and based on given Volume
}
\examples{
calc_kel_single_tdm(dose = 1000, t = 18)
}
