% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_t12.R
\name{calc_t12}
\alias{calc_t12}
\title{Calculate half-life based on two points}
\usage{
calc_t12(t1, t2, y1, y2)
}
\arguments{
\item{t1}{first sampling timepoint}

\item{t2}{second sampling timepoint}

\item{y1}{first sample value}

\item{y2}{second sample value}
}
\description{
based on two sampling points (in same interval)
}
\examples{
calc_t12(3, 24, 30, 10)
}
