% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_treosulfan_success.R
\name{calc_treosulfan_success}
\alias{calc_treosulfan_success}
\title{Calculate probability of therapeutic success for treosulfan}
\usage{
calc_treosulfan_success(cumulative_auc)
}
\arguments{
\item{cumulative_auc}{Cumulative AUC mg hour/L.}
}
\value{
Probability of therapeutic success for treosulfan.
}
\description{
Calculate probability of therapeutic success for treosulfan as a function of
cumulative AUC. This model was based on a data set of children receiving
treosulfan once daily for 3 days and may not correspond to other dosing
intervals or patient populations.
}
\note{
The parameter estimates used in this function match Figure 3a in the cited
paper, but differ slightly from the values provided in Table 2, which were
rounded. The decimal values used here were solved for using results cited in
the paper, and are close but not an exact match.
}
\examples{
calc_treosulfan_success(c(3863, 4800, 4829, 6037))

curve(
  calc_treosulfan_success,
  from = 2000, 
  to = 15000, 
  log = "x", 
  ylim = c(0, 1),
  xlab = "Cumulative AUC",
  ylab = "P(Success)"
)
}
\references{
\href{https://doi.org/10.1002/cpt.1715}{
Chiesa et al., Clinical Pharmacology and Therapeutics (2020)}
}
