% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covs_available.R
\name{check_covs_available}
\alias{check_covs_available}
\title{Returns parameters needed to calculate eGFR according to the method specified.}
\usage{
check_covs_available(cov_reqs = NULL, patient_covs = NULL,
  verbose = TRUE)
}
\arguments{
\item{cov_reqs}{vector of covariates required for calculating derived covariatiate}

\item{patient_covs}{named list of covariates}

\item{verbose}{stop and describe missing covariate(s)?}
}
\description{
Returns parameters needed to calculate eGFR according to the method specified.
}
\examples{
check_covs_available(egfr_cov_reqs('cockcroft_gault_ideal'), list(creat = 1, weight = 100, height = 160, sex = 'female', age = 90))
check_covs_available(egfr_cov_reqs('schwartz'), list(creat = c(0.2, 0.4), weight = 25, height = 80, sex = 'male', age = NULL))
}
