% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc2mol.R
\name{conc2mol}
\alias{conc2mol}
\title{Convert concentration to molar}
\usage{
conc2mol(conc = NULL, unit_conc = NULL, mol_weight = NULL, unit_mol = NULL)
}
\arguments{
\item{conc}{concentration in e.g. g/L}

\item{unit_conc, }{one of \code{g/l}, \code{mg/l}, \code{microg/l}, \verb{mcg/l", }ng/l\verb{, }mg/ml\verb{, }microg/ml\verb{, }mcg/ml\verb{, }ng/ml`}

\item{mol_weight}{concentration in g/mol}

\item{unit_mol}{one of \code{mol/L}, \code{mmol/mL}, \code{mmol/L}}
}
\description{
Convert concentration to molar
}
\examples{
conc2mol(100, unit_conc = "g/l", mol_weight = 180.15588)
}
