% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_age_unit.R
\name{convert_age_unit}
\alias{convert_age_unit}
\title{Convert age units between days, weeks, years}
\usage{
convert_age_unit(
  value,
  unit_in = c("days", "weeks", "years"),
  unit_out = c("days", "weeks", "years")
)
}
\arguments{
\item{value}{Value to convert}

\item{unit_in}{Unit of `value`}

\item{unit_out}{Output unit}
}
\value{
`value` converted to the appropriate unit
}
\description{
This function converts ages between units of days, weeks, and years. It does
not account for leap years.
}
\examples{
convert_age_unit(7, "days", "weeks")
convert_age_unit(365, "days", "years")
convert_age_unit(52, "weeks", "years")
convert_age_unit(25, "days", "weeks")
}
