% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_conc_unit.R
\name{convert_conc_unit}
\alias{convert_conc_unit}
\title{Convert concentration units}
\usage{
convert_conc_unit(value, unit_in, unit_out, mol_weight = NULL)
}
\arguments{
\item{value}{Value to convert}

\item{unit_in}{Input unit}

\item{unit_out}{Output unit}

\item{mol_weight}{Molecular weight in g/mol (required if converting to/from molar units)}
}
\description{
Lower-level function called by `convert_creat_unit()`,
`convert_albumin_unit()`, and others.
}
\seealso{
valid_units
}
\keyword{internal}
