% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_creat_unit.R
\name{convert_creat_unit}
\alias{convert_creat_unit}
\title{Convert creatinine to different unit}
\usage{
convert_creat_unit(value = NULL, unit_in = "mg/dL")
}
\arguments{
\item{value}{serum creatinine in either mg/dL or micromol/L}

\item{unit_in}{unit, either `mg/dL` or `micromol/L`}
}
\description{
Convert creatinine to different unit
}
\examples{
convert_creat_unit(1)
convert_creat_unit(88.42, unit_in = "micromol/l")
}
