% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flow_unit.R
\name{convert_flow_unit}
\alias{convert_flow_unit}
\title{Convert flow (e.g. clearance) from / to units}
\usage{
convert_flow_unit(value = NULL, from = "l", to = "ml", weight = NULL)
}
\arguments{
\item{value}{dose}

\item{from}{from dose unit}

\item{to}{from dose unit}

\item{weight}{for allowing per kg conversion}
}
\description{
Convert flow (e.g. clearance) from / to units
}
