% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose2auc.R
\name{dose2auc}
\alias{dose2auc}
\title{Convert dose to expected AUCtau or AUCt for 1 compartment linear PK model}
\usage{
dose2auc(dose, CL, V, t_auc = NA)
}
\arguments{
\item{dose}{dose amount}

\item{CL}{Clearance}

\item{V}{Volume of distribution}

\item{t_auc}{if AUCtau is not known but only AUCt, `t_auc` specifies time until which AUC_t is calculated to be able to calculate dose}
}
\description{
Convert dose to expected AUCtau or AUCt for 1 compartment linear PK model
}
