% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_cov_reqs.R
\name{egfr_cov_reqs}
\alias{egfr_cov_reqs}
\title{Returns parameters needed to calculate eGFR according to the method specified.}
\usage{
egfr_cov_reqs(method)
}
\arguments{
\item{method}{egfr calculation method}
}
\description{
Returns parameters needed to calculate eGFR according to the method specified.
}
\examples{
egfr_cov_reqs('schwartz_revised')
}
