% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flow_unit.R
\name{find_factor}
\alias{find_factor}
\title{Helper function to grab the conversion factor from an input unit and given list}
\usage{
find_factor(full_unit, units = NULL, prefix = "^")
}
\description{
Helper function to grab the conversion factor from an input unit and given list
}
