% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_dose.R
\name{find_nearest_dose}
\alias{find_nearest_dose}
\title{Generic function to calculate the dose nearest to a specific dose unit increment}
\usage{
find_nearest_dose(dose = NULL, increment = 250, type = "round")
}
\arguments{
\item{dose}{dose value}

\item{increment}{available increments of dose}

\item{type}{how to round, one of `round`, `floor`, or `ceiling`}
}
\description{
Generic function to calculate the dose nearest to a specific dose unit increment
}
\examples{
find_nearest_dose(573)
find_nearest_dose(573, increment = 50)
}
