% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_interval.R
\name{find_nearest_interval}
\alias{find_nearest_interval}
\title{Generic function to calculate the interval nearest to a possible dosing interval}
\usage{
find_nearest_interval(
  interval = NULL,
  possible = c(4, 6, 8, 12, 24, 36, 48),
  type = "absolute"
)
}
\arguments{
\item{interval}{dose value}

\item{possible}{available increments of dose}

\item{type}{pick either `nearest` absolute interval, or nearest `lower`, or nearest `higher` interval.}
}
\description{
Generic function to calculate the interval nearest to a possible dosing interval
}
\examples{
find_nearest_interval(19.7)
find_nearest_interval(19.7, c(6, 8, 12))
}
