% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth-charts}
\alias{growth-charts}
\alias{weight_for_age,}
\alias{height_for_age,}
\alias{bmi_for_age_children,}
\alias{weight_for_height_infants,}
\alias{weight_for_height_children}
\alias{weight_for_age}
\alias{height_for_age}
\alias{bmi_for_age_children}
\alias{weight_for_height_infants}
\title{CDC Growth Charts}
\format{
\subsection{\code{weight_for_age}}{

An object of class \code{data.frame} with 484 rows
and 14 columns.
}

\subsection{\code{height_for_age}}{

An object of class \code{data.frame} with 484 rows
and 14 columns.
}

\subsection{\code{bmi_for_age_children}}{

An object of class \code{data.frame} with 438 rows
and 15 columns.
}

\subsection{\code{weight_for_height_infants}}{

An object of class \code{data.frame} with 120 rows
and 14 columns.
}

\subsection{\code{weight_for_height_children}}{

An object of class \code{data.frame} with 92 rows
and 15 columns.
}
}
\source{
National Center for Health Statistics. (n.d.). CDC Growth Charts Data Files.
\url{https://www.cdc.gov/growthcharts/cdc-growth-charts.htm}
}
\usage{
weight_for_age

height_for_age

bmi_for_age_children

weight_for_height_infants

weight_for_height_children
}
\description{
Data used to produce the United States Growth Charts smoothed percentile
curves for infants and older children.
}
\details{
These data sets contain the L (power in the Box-Cox transformation), M
(median), and S (generalized coefficient of variation) parameters needed to
generate exact percentiles and z-scores for seven different growth charts,
along with the 3rd, 5th, 10th, 25th, 50th, 75th, 90th, 95th, and 97th
percentile values by sex (1 = male; 2 = female) and single month of age
(listed at the half-month point for the entire month).
}
\seealso{
Functions to calculate growth metrics: \link{pct_weight_for_age},
\link{pct_height_for_age}, \link{pct_bmi_for_age}, \link{pct_weight_for_height}

LMS growth chart equations: \link{lms}
}
\keyword{datasets}
