% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ibw.R
\name{ibw_standard}
\alias{ibw_standard}
\title{Calculate IBW using "standard" equation}
\usage{
ibw_standard(age, height = NULL, sex = NULL)
}
\arguments{
\item{age}{age in years}

\item{height}{height in cm}

\item{sex}{sex}
}
\description{
Calculate IBW using "standard" equation
}
