% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aki_stage.R
\name{kdigo_stage}
\alias{kdigo_stage}
\title{Calculate AKI stage according to KDIGO criteria}
\usage{
kdigo_stage(dat, baseline_scr, age)
}
\arguments{
\item{dat}{Data frame containing at least the following columns:
\itemize{
\item \code{scr}: serum creatinine
\item \code{t}: creatinine sample times in hours
\item \code{baseline_scr_diff}: difference between baseline scr and scr at current
timepoint
\item \code{egfr}: eGFR at timepoint
}}

\item{baseline_scr}{Baseline serum creatinine value (numeric)}

\item{age}{Patient age}
}
\description{
Calculate AKI stage according to KDIGO criteria
}
\keyword{internal}
