% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{lms}
\alias{lms}
\alias{lms_for_x,}
\alias{lms_for_z}
\alias{lms_for_x}
\title{LMS growth chart equations}
\source{
National Center for Health Statistics. (n.d.). CDC Growth Charts Data Files.
\url{https://www.cdc.gov/growthcharts/cdc-growth-charts.htm}
}
\usage{
lms_for_x(l, m, s, z = NULL, p = NULL)

lms_for_z(l, m, s, x, value = c("z", "p"))
}
\arguments{
\item{l}{The power in the Box-Cox transformation (L).}

\item{m}{The median (M).}

\item{s}{The generalized coefficient of variation (S)}

\item{z, p}{The z-score (Z) or percentile to return the value of a given
physical measurement (X) at. Only one of \code{z} or \code{p} can be provided, not
both.}

\item{x}{A given physical measurement (X) to return the corresponding z-score
(Z) or percentile for.}

\item{value}{A character string determining whether to return the
corresponding z-score (Z) or percentile.}
}
\value{
\code{lms_for_x()} returns the value of a given physical measurement (X)
at a particular z-score (Z) or percentile. \code{lms_for_z()} returns the
z-score (Z) or corresponding percentile for a given measurement (X).
}
\description{
LMS equations to calculate z-scores, percentiles, and other metrics for
infants and children.
}
\examples{
lms_for_x(l = -0.1600954, m = 9.476500305, s = 0.11218624, z = -1.645)

lms_for_z(
  l = -0.1600954, m = 9.476500305, s = 0.11218624, x = 9.7, value = "p"
)
}
\seealso{
CDC Growth Charts data with LMS parameters: \link{weight_for_age},
\link{height_for_age}, \link{bmi_for_age_children}, \link{weight_for_height_infants},
\link{weight_for_height_children}

Functions to calculate growth metrics: \link{pct_weight_for_age}, \link{pct_height_for_age},
\link{pct_bmi_for_age}, \link{pct_weight_for_height}
}
