% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_linear.R
\name{mean_step}
\alias{mean_step}
\title{Calculate mean between two values in a vector
Used in NCA for calculation of mean between two or more timepoints.}
\usage{
mean_step(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
a vector of length(x)-1 with the mean timepoints
}
\description{
Calculate mean between two values in a vector
Used in NCA for calculation of mean between two or more timepoints.
}
\keyword{internal}
