% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_x_for_y.R
\name{median_weight_for_age}
\alias{median_weight_for_age}
\alias{median_height_for_age}
\alias{median_bmi_for_age}
\alias{median_weight_for_height}
\title{Median growth values for infants and children}
\usage{
median_weight_for_age(
  age,
  age_units = c("years", "months", "weeks", "days"),
  sex
)

median_height_for_age(
  age,
  age_units = c("years", "months", "weeks", "days"),
  sex
)

median_bmi_for_age(age, age_units = c("years", "months", "weeks", "days"), sex)

median_weight_for_height(
  height,
  height_units = c("centimetres", "metres", "feet", "inches"),
  sex,
  population = c("infants", "children")
)
}
\arguments{
\item{age}{A numeric vector of ages in the unit specified in \code{age_units}.}

\item{age_units}{A character string specifying the units of all \code{age} values.}

\item{sex}{A character vector specifying patient sex. Either \code{"male"} or
\code{"female"}.}

\item{height}{A numeric vector of heights in the unit specified in \code{height_units}.}

\item{height_units}{A character string specifying the units of all \code{height} values.}

\item{population}{A character string specifying the population table to use
for \code{\link[=median_weight_for_height]{median_weight_for_height()}}. Either "infants" (birth to 36 months) or
"children" (2 to 20 years).}
}
\value{
A numeric vector of median weight, height, or BMI values for the
given ages/heights.
}
\description{
Calculate median weight, height, and BMI for age/height for infants and
children using CDC Growth Charts data and equations.
}
\examples{
median_weight_for_age(3.5, "months", sex = "male")
}
\seealso{
Functions to calculate growth metrics: \link{pct_weight_for_age},
\link{pct_height_for_age}, \link{pct_bmi_for_age}, \link{pct_weight_for_height}

CDC Growth Charts data: \link{weight_for_age}, \link{height_for_age},
\link{bmi_for_age_children}, \link{weight_for_height_infants},
\link{weight_for_height_children}
}
