% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol2conc.R
\name{mol2conc}
\alias{mol2conc}
\title{Convert molar to concentration}
\usage{
mol2conc(mol = NULL, unit_mol = NULL, unit_conc = NULL, mol_weight = NULL)
}
\arguments{
\item{mol}{concentration in molars}

\item{unit_mol}{unit of input concentration (molar), one of \code{mol/L}, \code{mmol/mL}, \code{mmol/L}}

\item{unit_conc, }{output unit, one of \code{g/l}, \code{mg/l}, \code{microg/l}, \verb{mcg/l", }ng/l\verb{, }mg/ml\verb{, }microg/ml\verb{, }mcg/ml\verb{, }ng/ml`}

\item{mol_weight}{concentration in g/mol}
}
\description{
Convert molar to concentration
}
\examples{
mol2conc(1, unit_mol = "mmol/l", mol_weight = 180)
}
