% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_linear.R
\name{nca_linear}
\alias{nca_linear}
\title{NCA core function to perform linear trapezoid calculations for post-infusion}
\usage{
nca_linear(data)
}
\arguments{
\item{data}{dataset passed from nca() with `time` and `dv` columns}
}
\description{
NCA core function to perform linear trapezoid calculations for post-infusion
}
\keyword{internal}
