% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_trapezoid.R
\name{nca_trapezoid}
\alias{nca_trapezoid}
\title{NCA core function to perform log-linear trapezoid calculations for post-infusion}
\usage{
nca_trapezoid(data)
}
\arguments{
\item{data}{dataset passed from nca() with `time` and `dv` columns}
}
\description{
NCA core function to perform log-linear trapezoid calculations for post-infusion
}
\keyword{internal}
