% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_for_age_generic.R
\name{pct_for_age_generic}
\alias{pct_for_age_generic}
\title{Percentile height or weight for age for children}
\usage{
pct_for_age_generic(
  age = NULL,
  value = NULL,
  sex = NULL,
  variable = "weight",
  ...
)
}
\arguments{
\item{age}{age in years}

\item{value}{height in kg. Optional, if specified, will calculate closest percentile and return in list as `percentile`}

\item{sex}{either `male` or `female`}

\item{variable}{weight or height?}

\item{...}{parameters passed to `read_who_table()`}
}
\description{
This is the underlying function, the exposed functions are pct_weight_for_age() and pct_height_for_age()
Based on tables from WHO: http://www.who.int/childgrowth/standards/height_for_age/en/
}
\keyword{internal}
