% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_weight_for_age.R
\name{pct_weight_for_age}
\alias{pct_weight_for_age}
\title{Percentile weight for age for children}
\usage{
pct_weight_for_age(age = NULL, weight = NULL, sex = NULL, ...)
}
\arguments{
\item{age}{age in years}

\item{weight}{weight in kg. Optional, if specified, will calculate closest percentile and return in list as `percentile`}

\item{sex}{either `male` or `female`}

\item{...}{parameters passed to `read_who_table()`}
}
\description{
Based on tables from WHO: http://www.who.int/childgrowth/standards/weight_for_age/en/
}
