% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_1cmt_bolus_dose_from_cmax.R
\name{pk_1cmt_bolus_dose_from_cmax}
\alias{pk_1cmt_bolus_dose_from_cmax}
\title{Calculate dose to achieve steady state Cmax
for 1-compartmental PK model bolus dosing at steady state}
\usage{
pk_1cmt_bolus_dose_from_cmax(cmax = 1, tau = 12, CL = 3, V = 30)
}
\arguments{
\item{cmax}{desired trough concentration}

\item{tau}{dosing interval}

\item{CL}{clearance}

\item{V}{volume of distribution}
}
\description{
Calculate dose to achieve steady state Cmax
for 1-compartmental PK model bolus dosing at steady state
}
\examples{
dos <- pk_1cmt_bolus_dose_from_cmax(
  cmax = 10, tau = 12, CL = 5, V = 50)
find_nearest_dose(dos, 100)
}
