% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_1cmt_inf_cmax_ss.R
\name{pk_1cmt_inf_cmax_ss}
\alias{pk_1cmt_inf_cmax_ss}
\title{Cmax for linear 1-compartment PK model at steady state}
\usage{
pk_1cmt_inf_cmax_ss(dose, tau, CL, V, t_inf, ruv = NULL)
}
\arguments{
\item{dose}{dose}

\item{tau}{dosing interval}

\item{CL}{clearance}

\item{V}{volume of distrubition}

\item{t_inf}{infusion time}

\item{ruv}{residual variability, specified as list with optional arguments for proportional, additive, or exponential components, e.g. `list(prop=0.1, add=1, exp=0)`}
}
\description{
Takes single values for dose or model parameters, or vector of either dose or parameters (but not both).
}
\examples{
pk_1cmt_inf_cmax_ss(dose = 500, tau = 12, t_inf = 2, CL = 5, V = 50)
}
