% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_1cmt_t12.R
\name{pk_1cmt_t12}
\alias{pk_1cmt_t12}
\title{Calculate terminal half-life for 1-compartment model}
\usage{
pk_1cmt_t12(CL = 3, V = 30)
}
\arguments{
\item{CL}{clearance}

\item{V}{volume of central compartment}
}
\description{
Calculate terminal half-life for 1-compartment model
}
\examples{
pk_1cmt_t12(CL = 5, V = 50)

}
