% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_2cmt_inf.R
\name{pk_2cmt_inf}
\alias{pk_2cmt_inf}
\title{Concentration predictions for 2-compartmental PK model, single or multiple infusions}
\usage{
pk_2cmt_inf(
  t = c(0:24),
  dose = 100,
  tau = 12,
  t_inf = 1,
  CL = 3,
  V = 30,
  Q = 2,
  V2 = 20,
  ruv = NULL
)
}
\arguments{
\item{t}{vector of time}

\item{dose}{dose}

\item{tau}{dosing interval}

\item{t_inf}{infusion time}

\item{CL}{clearance}

\item{V}{volume of central compartment}

\item{Q}{inter-compartimental clearance}

\item{V2}{volume of peripheral compartment}

\item{ruv}{residual error (list)}
}
\description{
Concentration predictions for 2-compartmental PK model, single or multiple infusions
}
