% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_2cmt_inf_cmin_ss.R
\name{pk_2cmt_inf_cmin_ss}
\alias{pk_2cmt_inf_cmin_ss}
\title{Cmin (trough) for 2-compartmental PK model, bolus dosing at steady state}
\usage{
pk_2cmt_inf_cmin_ss(
  dose = 100,
  tau = 12,
  t_inf = 1,
  CL = 3,
  V = 30,
  Q = 2,
  V2 = 20,
  ruv = NULL
)
}
\arguments{
\item{dose}{dose}

\item{tau}{dosing interval}

\item{t_inf}{infusion time}

\item{CL}{clearance}

\item{V}{volume of central compartment}

\item{Q}{inter-compartimental clearance}

\item{V2}{volume of peripheral compartment}

\item{ruv}{residual error (list)}
}
\description{
Cmin (trough) for 2-compartmental PK model, bolus dosing at steady state
}
\examples{
pk_2cmt_inf_cmin_ss(
  dose = 1000, tau = 12, t_inf = 2,
  CL = 5, V = 50, Q = 20, V2 = 200)
}
