% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_2cmt_t12.R
\name{pk_2cmt_t12}
\alias{pk_2cmt_t12}
\title{Calculate half-life(s) for 2-compartment model}
\usage{
pk_2cmt_t12(CL = 3, V = 30, Q = 2, V2 = 20, phase = c("both", "alpha", "beta"))
}
\arguments{
\item{CL}{clearance}

\item{V}{volume of central compartment}

\item{Q}{inter-compartimental clearance}

\item{V2}{volume of peripheral compartment}

\item{phase}{`alpha`, `beta` (default) or `both` to indicate initial (distribution) or terminal (elimination) phase.}
}
\description{
Calculate half-life(s) for 2-compartment model
}
\examples{
pk_2cmt_t12(CL = 5, V = 50, Q = 20, V2 = 200)
}
