% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_2cmt_t12_interval.R
\name{pk_2cmt_t12_interval}
\alias{pk_2cmt_t12_interval}
\title{Calculate average half-life for 2-compartment model during a specific interval}
\usage{
pk_2cmt_t12_interval(CL = 3, V = 30, Q = 2, V2 = 20, tau = 12, t_inf = NULL)
}
\arguments{
\item{CL}{clearance}

\item{V}{volume of central compartment}

\item{Q}{inter-compartimental clearance}

\item{V2}{volume of peripheral compartment}

\item{tau}{interval (hours)}

\item{t_inf}{infusion time (hours)}
}
\description{
Calculate average half-life for 2-compartment model during a specific interval
}
\examples{
pk_2cmt_t12_interval(CL = 5, V = 50, Q = 20, V2 = 200, tau = 12, t_inf = 2)

}
