% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nca_output.R
\name{print.nca_output}
\alias{print.nca_output}
\title{Print function for output from nca()}
\usage{
\method{print}{nca_output}(x, ...)
}
\arguments{
\item{x}{output object (list) from nca()}

\item{...}{variables past on to print function}
}
\description{
Print function for output from nca()
}
\keyword{internal}
