% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_who_table.R
\name{read_who_table}
\alias{read_who_table}
\title{Internal function to read WHO growth tables from package or download from WHO}
\usage{
read_who_table(sex = NULL, age = NULL, type = "wfa",
  who_url = "http://www.who.int/entity/childgrowth/standards",
  download = FALSE)
}
\arguments{
\item{sex, }{either `male` or `female`}

\item{age}{age in years}

\item{type}{table type, choose from `wfa` (weight for age), `lhfa` (length for age)}

\item{who_url}{base URL for WHO growth tables}

\item{download}{download current tables from WHO?}
}
\description{
Internal function to read WHO growth tables from package or download from WHO
}
