% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_lt_gt}
\alias{remove_lt_gt}
\title{Remove less-than/greater-than symbols and convert to numeric}
\usage{
remove_lt_gt(x)
}
\arguments{
\item{x}{Vector of numbers possibly containing extraneous strings.}
}
\value{
If non-numeric characters were successfully removed, returns a
  numeric vector. If some elements of `x` contained other characters, their
  original value will be returned and the result will be a character vector.
}
\description{
The following characters will be removed from strings: <, >, =, space. If
string contains other characters, the original string will be returned.
}
\keyword{internal}
