% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight2kg.R
\name{weight2kg}
\alias{weight2kg}
\title{Convert any weight unit to kg}
\usage{
weight2kg(value = NULL, unit = NULL)
}
\arguments{
\item{value}{weight in any allowed unit}

\item{unit}{unit of weight, one of "lbs", "pound", "pounds", "oz", "ounce", "ounces"}
}
\description{
Convert any weight unit to kg
}
\examples{
weight2kg(250, unit = "oz")
weight2kg(250, unit = "pounds")
weight2kg(250, unit = "lbs")
}
