% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ruv.R
\name{add_ruv}
\alias{add_ruv}
\title{Add residual variability to data}
\usage{
add_ruv(x, ruv = list())
}
\arguments{
\item{x}{data}

\item{ruv}{list with arguments prop, add, exp}
}
\description{
Add residual variability to data
}
\examples{
y <- pk_1cmt_inf()$y
y + add_ruv(y, list(prop = 0.1, add = 0.05))
}
