% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aki_stage.R
\name{calc_aki_stage}
\alias{calc_aki_stage}
\title{Calculate AKI stage}
\usage{
calc_aki_stage(scr = NULL, times = NULL, method = "kDIGO",
  baseline_scr = NULL, baseline_egfr = NULL, age = NULL, egfr = NULL,
  egfr_method = "cockcroft_gault", force_numeric = FALSE,
  recursive = TRUE, verbose = TRUE, return_object = TRUE, ...)
}
\arguments{
\item{scr}{serum creatinine in mg/dL. Use `convert_creat()` to convert from mmol/L.}

\item{times}{creatinine sample times in hours}

\item{method}{classification method, one of `KDIGO`, `AKIN`, `RIFLE`, `pRIFLE`}

\item{baseline_scr}{baseline serum creatinine, required for `RIFLE` classifation. Will take median of `scr` values if `NULL`.}

\item{baseline_egfr}{baseline eGFR, required for `AKIN`, `RIFLE`, and `pRIFLE` classifations. Will take median of `egfr` values if `NULL`.}

\item{age}{age in years, needed when eGFR is used in the classification method}

\item{egfr}{eGFR in ml/min/1.73m^2. Optional, can also be calcualted if `age`, `weight`, `height`, `sex`, `egfr_method` are specified as arguments.}

\item{egfr_method}{eGFR calculation method, used by `calc_egfr()`}

\item{force_numeric}{keep stage numeric (1, 2, or 3), instead of e.g. "R", "I", "F" as in RIFLE. Default `FALSE`.}

\item{recursive}{option for KDIGO classification method only. Use recursive calculation (if `FALSE` will only take last observation into account)}

\item{verbose}{verbose (`TRUE` or `FALSE`)}

\item{return_object}{return object with detailed data (default `TRUE`). If `FALSE`, will just return maximum stage.}

\item{...}{arguments passed on to `calc_egfr()`}
}
\description{
Calculate AKI class based on serum creatinine values over time, using various
methods for children (pRIFLE) and adults (RIFLE, kDIGO)
}
\examples{
calc_aki_stage(
  scr = c(0.7, 0.9, 1.8, 1.5),
  t = c(0, 40, 100, 130),
  age = 50, weight = 60,
  height = 170, sex = "female")

}
