% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aki_stage.R
\name{calc_aki_stage}
\alias{calc_aki_stage}
\title{Calculate AKI stage}
\usage{
calc_aki_stage(
  scr = NULL,
  times = NULL,
  method = "kdigo",
  baseline_scr = "median",
  baseline_egfr = 120,
  age = NULL,
  egfr = NULL,
  egfr_method = NULL,
  force_numeric = FALSE,
  override_prifle_baseline = FALSE,
  recursive = TRUE,
  verbose = TRUE,
  return_object = TRUE,
  ...
)
}
\arguments{
\item{scr}{serum creatinine in mg/dL. Use `convert_creat()` to convert from
mmol/L. Values below the detection limit ("<0.2") will be converted to
numeric (0.2)}

\item{times}{creatinine sample times in hours}

\item{method}{classification method, one of `KDIGO`, `RIFLE`, `pRIFLE` (case insensitive)}

\item{baseline_scr}{baseline serum creatinine, required for `RIFLE` classifation. Will use value if numeric. If `character`, can be either `median` or `expected`. The latter will use the expected value based on sex and age.}

\item{baseline_egfr}{baseline eGFR, required for `RIFLE` classifations. Will take median of `egfr` values if `NULL`.}

\item{age}{age in years, needed when eGFR is used in the classification method}

\item{egfr}{eGFR in ml/min/1.73m^2. Optional, can also be calcualted if `age`, `weight`, `height`, `sex`, `egfr_method` are specified as arguments.}

\item{egfr_method}{eGFR calculation method, used by `calc_egfr()`. If NULL, will pick default based on classification system (`cockroft_gault` for RIFLE / kDIGO, `revised_schwartz` for pRIFLE).}

\item{force_numeric}{keep stage numeric (1, 2, or 3), instead of e.g. "R", "I", "F" as in RIFLE. Default `FALSE`.}

\item{override_prifle_baseline}{by default, `pRIFLE` compares eGFR to 120 ml/min. Override by setting to TRUE.}

\item{recursive}{option for KDIGO classification method only. Use recursive calculation (if `FALSE` will only take last observation into account)}

\item{verbose}{verbose (`TRUE` or `FALSE`)}

\item{return_object}{return object with detailed data (default `TRUE`). If `FALSE`, will just return maximum stage.}

\item{...}{arguments passed on to `calc_egfr()`}
}
\description{
Calculate AKI class based on serum creatinine values over time, using various
methods for children (pRIFLE) and adults (RIFLE, kDIGO)
}
\examples{
calc_aki_stage(
  scr = c(0.7, 0.9, 1.8, 1.5),
  t = c(0, 40, 100, 130),
  age = 50, weight = 60,
  height = 170, sex = "female")

}
