% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bsa.R
\name{calc_bsa}
\alias{calc_bsa}
\title{Calculate body surface area}
\usage{
calc_bsa(
  weight = NULL,
  height = NULL,
  method = c("dubois", "mosteller", "haycock", "gehan_george", "boyd")
)
}
\arguments{
\item{weight}{weight}

\item{height}{height}

\item{method}{estimation method, choose from `dubois`, `mosteller`, `haycock`, `gehan_george`, `boyd`}
}
\value{
Returns a list of the following elements:
\item{value}{Body Surface Area (BSA) in units of m2}
\item{unit}{Unit describing BSA, (m2)}
}
\description{
Get an estimate of body-surface area (in m2) based on weight and height
}
\examples{
calc_bsa(weight = 70, height = 170)
calc_bsa(weight = 70, height = 170, method = "gehan_george")
}
