% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cefepime_neurotoxicity.R
\name{calc_cefepime_neurotoxicity}
\alias{calc_cefepime_neurotoxicity}
\title{Calculate Cefepime-associated Neurotoxicity}
\usage{
calc_cefepime_neurotoxicity(trough_concentration)
}
\arguments{
\item{trough_concentration}{Cefepime plasma trough concentration (mg/L).}
}
\value{
Probability of cefepime-associated neurotoxicity.
}
\description{
Calculate probability of cefepime-associated neurotoxicity as a function of
cefepime plasma trough concentrations.
}
\examples{
calc_cefepime_neurotoxicity(trough_concentration = c(12, 16))
}
\references{
\href{https://doi.org/10.1016/j.cmi.2019.06.028}{
Boschung-Pasquier et al., Clinical Microbiology and Infection (2020)}
}
