% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egfr_cystatin.R
\name{calc_egfr_cystatin}
\alias{calc_egfr_cystatin}
\title{Calculate eGFR based on Cystatin C measurements}
\usage{
calc_egfr_cystatin(cystatin = NULL, cystatin_unit = "mg/L",
  method = "grubb", unit_out = "mL/min")
}
\arguments{
\item{cystatin}{serum cystatin concentration (mg/L)}

\item{cystatin_unit, }{only `mg/L` available}

\item{method}{eGFR estimation method, choose from `grubb`, `larsson`}

\item{unit_out}{eGFR output unit, choose from `ml/min`, `ml/hr`, `l/min`, `l/hr`}
}
\description{
Calculate eGFR based on Cystatin C measurements
}
