% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc2mol.R
\name{conc2mol}
\alias{conc2mol}
\title{Convert concentration to molar}
\usage{
conc2mol(conc = NULL, unit_conc = NULL, mol_weight = NULL, unit_mol = NULL)
}
\arguments{
\item{conc}{concentration in e.g. g/L}

\item{unit_conc, }{one of `g/l`, `mg/l`, `microg/l`, `mcg/l", `ng/l`, `mg/ml`, `microg/ml`, `mcg/ml`, `ng/ml`}

\item{mol_weight}{concentration in g/mol}

\item{unit_mol}{one of `mol/L`, `mmol/mL`, `mmol/L`}
}
\description{
Convert concentration to molar
}
\examples{
conc2mol(100, unit_conc = "g/l", mol_weight = 180.15588)
}
