% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flow_unit.R
\name{find_factor}
\alias{find_factor}
\title{Helper function to grab the conversion factor from an input unit and given list}
\usage{
find_factor(full_unit, units = NULL, prefix = "^")
}
\arguments{
\item{full_unit}{full unit, e.g. "mL/min/kg"}

\item{units}{unit specification list, e.g. `list("ml" = 1/1000, "dl" = 1/10, "l" = 1)`}

\item{prefix}{prefix used in matching units, e.g. "^" only matches at start of string while "_" matches units specified as "/"}
}
\description{
Helper function to grab the conversion factor from an input unit and given list
}
\keyword{internal}
