% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol2conc.R
\name{mol2conc}
\alias{mol2conc}
\title{Convert molar to concentration}
\usage{
mol2conc(mol = NULL, unit_mol = NULL, unit_conc = NULL, mol_weight = NULL)
}
\arguments{
\item{mol}{concentration in molars}

\item{unit_mol}{unit of input concentration (molar), one of `mol/L`, `mmol/mL`, `mmol/L`}

\item{unit_conc, }{output unit, one of `g/l`, `mg/l`, `microg/l`, `mcg/l", `ng/l`, `mg/ml`, `microg/ml`, `mcg/ml`, `ng/ml`}

\item{mol_weight}{concentration in g/mol}
}
\description{
Convert molar to concentration
}
\examples{
mol2conc(1, unit_mol = "mmol/l", mol_weight = 180)
}
