% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_bmi_for_age.R
\name{pct_bmi_for_age}
\alias{pct_bmi_for_age}
\title{Percentile BMI for age for children}
\usage{
pct_bmi_for_age(
  age = NULL,
  bmi = NULL,
  sex = NULL,
  height = NULL,
  return_median = FALSE,
  ...
)
}
\arguments{
\item{age}{age in years}

\item{bmi}{bmi Optional, if specified, will calculate closest percentile and return in list as `percentile`}

\item{sex}{either `male` or `female`}

\item{height}{height}

\item{return_median}{just return the median expected value}

\item{...}{parameters passed to `read_who_table()`}
}
\description{
Based on tables from WHO: http://www.who.int/growthref/who2007_bmi_for_age/en/
}
\examples{
pct_bmi_for_age(age = 8, sex = "male")
pct_bmi_for_age(age = 8, bmi = 15, sex = "male")
}
