% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_who_table.R
\name{read_who_table}
\alias{read_who_table}
\title{Read WHO growth tables}
\usage{
read_who_table(sex = NULL, age = NULL, type = "wfa")
}
\arguments{
\item{sex, }{either \code{male} or \code{female}}

\item{age}{age in years}

\item{type}{table type, choose from \code{wfa} (weight for age), \code{lhfa} (length for age)}
}
\description{
Provides a data frame of the WHO growth table for a given age, sex, and type
of measurement.
}
\details{
This function uses files included in \code{system.file(package = "clinPK")}.
Previously this function also gave the option to download the tables from
WHO, but the original URL ("http://www.who.int/entity/childgrowth/standards")
no longer exists as of 2021-05-19.
}
